app/requirements.txt
openai
streamlit>=1.28.0
python-dotenv==1.0.0
PyYAML
langchain==0.2.17
langchain-community==0.2.5
pydantic==2.10.3
pydantic-core==2.27.1
pandas>=1.5.0
python-docx
chromadb==0.5.15
urllib3==2.0.4
PyMuPDF==1.23.5
tabulate==0.9.0
tqdm==4.66.1
fastapi>=0.111.0
uvicorn[standard]
python-multipart
langdetect==1.0.9
PyJWT
prometheus-client
plotly>=5.0.0
llama-parse>=0.4.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
moviepy>=1.0.3
python-magic>=0.4.27
nltk>=3.8
unstructured>=0.10.0
GitPython>=3.1.40
aiohttp>=3.9.0
fnmatch2>=0.0.8
langchain-core==0.2.43
langchain-text-splitters==0.2.4
